# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Aquaventure Swim School API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /contacts:
    post:
      operationId: submitContact
      tags: [contacts]
      summary: Submit a contact inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitContactBody"
      responses:
        "201":
          description: Contact submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/faq:
    get:
      operationId: listAdminFaq
      tags: [admin]
      summary: List all FAQ entries (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: FAQ list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FaqEntry"
    post:
      operationId: createFaq
      tags: [admin]
      summary: Create a FAQ entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFaqBody"
      responses:
        "201":
          description: Created FAQ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqEntry"
  /admin/faq/{id}:
    patch:
      operationId: updateFaq
      tags: [admin]
      summary: Update a FAQ entry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFaqBody"
      responses:
        "200":
          description: Updated FAQ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqEntry"
    delete:
      operationId: deleteFaq
      tags: [admin]
      summary: Delete a FAQ entry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/contacts:
    get:
      operationId: listContacts
      tags: [admin]
      summary: List all contact inquiries
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
  /admin/contacts/{id}/status:
    patch:
      operationId: updateContactStatus
      tags: [admin]
      summary: Update contact status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactStatusBody"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
...
  /swimmers/my:
    get:
      operationId: getMySwimmerProfiles
      tags: [swimmers]
      summary: Get swimmer profiles for the authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Swimmer profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SwimmerProfile"
  /swimmers/{id}:
    patch:
      operationId: updateSwimmerProfile
      tags: [swimmers]
      summary: Create or update a swimmer profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertSwimmerProfileBody"
      responses:
        "200":
          description: Updated swimmer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwimmerProfile"
  /swimmers:
    post:
      operationId: createSwimmerProfile
      tags: [swimmers]
      summary: Create a swimmer profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertSwimmerProfileBody"
      responses:
        "201":
          description: Created swimmer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwimmerProfile"
...
  /progress/my:
    get:
      operationId: getMyProgress
      tags: [progress]
      summary: Get skill progress for the authenticated user's enrollments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Progress records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressRecord"
  /admin/progress:
    get:
      operationId: listProgress
      tags: [admin]
      summary: List skill progress (optionally filtered by enrollment)
      security:
        - bearerAuth: []
      parameters:
        - name: enrollmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Progress records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressRecord"
    post:
      operationId: logProgress
      tags: [admin]
      summary: Log a skill achievement for an enrollment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogProgressBody"
      responses:
        "201":
          description: Progress logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressRecord"
  /admin/progress/{id}:
    delete:
      operationId: deleteProgress
      tags: [admin]
      summary: Delete a skill progress record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressRecord"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, parent, swimmer]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    SwimmerProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        swimmerName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        level:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - swimmerName
        - createdAt
        - updatedAt
    UpsertSwimmerProfileBody:
      type: object
      properties:
        swimmerName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        level:
          type: ["string", "null"]
      required:
        - swimmerName
...
    FaqEntry:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        answer:
          type: string
        category:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - question
        - answer
        - category
        - sortOrder
        - createdAt
    CreateFaqBody:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        category:
          type: string
        sortOrder:
          type: integer
      required:
        - question
        - answer
        - category
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [unread, read, replied]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - subject
        - message
        - status
        - createdAt
    SubmitContactBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string
      required:
        - name
        - email
        - subject
        - message
    UpdateContactStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [unread, read, replied]
      required:
        - status
...
    ProgressRecord:
      type: object
      properties:
        id:
          type: integer
        enrollmentId:
          type: integer
        skill:
          type: string
        achievedAt:
          type: string
        notes:
          type: ["string", "null"]
        swimmerName:
          type: ["string", "null"]
        programName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - enrollmentId
        - skill
        - achievedAt
        - createdAt
    LogProgressBody:
      type: object
      properties:
        enrollmentId:
          type: integer
        skill:
          type: string
        achievedAt:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - enrollmentId
        - skill
        - achievedAt
...